use string resources. add license header. cleanup
