+++ app/src/main/java/com/auth0/android/lock/app/DemoActivity.java
@@ -1,3 +1,27 @@
+/*
+ * DemoActivity.java
+ *
+ * Copyright (c) 2016 Auth0 (http://auth0.com)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
 package com.auth0.android.lock.app;
 
 import android.content.Intent;
@@ -6,17 +30,14 @@ import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 
 import com.auth0.Auth0;
-import com.auth0.Auth0Exception;
 import com.auth0.android.lock.AuthenticationCallback;
 import com.auth0.android.lock.Lock;
+import com.auth0.android.lock.utils.LockException;
 import com.auth0.authentication.result.Authentication;
 
-/**
- * Created by lbalmaceda on 1/21/16.
- */
 public class DemoActivity extends AppCompatActivity implements AuthenticationCallback {
     private static final String AUTH0_CLIENT_ID = "Owu62gnGsRYhk1v9SfB3c6IUbIJcRIze";
-    private static final String AUTH0_DOMAIN = "http://lbalmaceda.auth0.com";
+    private static final String AUTH0_DOMAIN = "lbalmaceda.auth0.com";
     private static final int AUTH_REQUEST = 333;
     private Lock lock;
 
@@ -67,7 +88,7 @@ public class DemoActivity extends AppCompatActivity implements AuthenticationCal
     }
 
     @Override
-    public void onError(Auth0Exception error) {
+    public void onError(LockException error) {
 
     }
 }
