+++ lock/src/main/java/com/auth0/android/lock/Lock.java
@@ -1,3 +1,27 @@
+/*
+ * Lock.java
+ *
+ * Copyright (c) 2016 Auth0 (http://auth0.com)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
 package com.auth0.android.lock;
 
 import android.app.Activity;
@@ -8,7 +32,7 @@ import android.content.IntentFilter;
 import android.support.v4.content.LocalBroadcastManager;
 
 import com.auth0.Auth0;
-import com.auth0.Auth0Exception;
+import com.auth0.android.lock.utils.LockException;
 import com.auth0.authentication.result.Authentication;
 import com.auth0.authentication.result.Token;
 import com.auth0.authentication.result.UserProfile;
@@ -17,9 +41,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-/**
- * Created by nikolaseu on 1/21/16.
- */
 public class Lock {
 
     private AuthenticationCallback callback;
@@ -127,7 +148,7 @@ public class Lock {
         if (idToken != null && accessToken != null) {
             callback.onAuthentication(authentication);
         } else {
-            Auth0Exception up = new Auth0Exception("Error parsing Authentication data");
+            LockException up = new LockException(R.string.com_auth0_social_error_authentication);
             callback.onError(up);
             //throw up. haha
         }
