+++ lock/src/main/java/com/auth0/android/lock/LockActivity.java
@@ -1,3 +1,27 @@
+/*
+ * LockActivity.java
+ *
+ * Copyright (c) 2016 Auth0 (http://auth0.com)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
 package com.auth0.android.lock;
 
 
@@ -14,9 +38,9 @@ import android.widget.FrameLayout;
 
 import com.auth0.Auth0Exception;
 import com.auth0.android.lock.events.SocialConnectionEvent;
-import com.auth0.android.lock.net.CallbackParser;
-import com.auth0.android.lock.net.IdentityProviderCallback;
-import com.auth0.android.lock.net.WebIdentityProvider;
+import com.auth0.android.lock.provider.CallbackParser;
+import com.auth0.android.lock.provider.IdentityProviderCallback;
+import com.auth0.android.lock.provider.WebIdentityProvider;
 import com.auth0.android.lock.utils.Application;
 import com.auth0.android.lock.utils.Configuration;
 import com.auth0.android.lock.views.LockProgress;
@@ -35,9 +59,6 @@ import org.json.JSONTokener;
 
 import java.io.IOException;
 
-/**
- * Created by lbalmaceda on 1/21/16.
- */
 public class LockActivity extends AppCompatActivity {
 
     private static final String TAG = LockActivity.class.getSimpleName();
@@ -92,9 +113,9 @@ public class LockActivity extends AppCompatActivity {
         lockBus.register(this);
         handler = new Handler(getMainLooper());
 
-        setContentView(R.layout.com_auth0_lock_activity_lock);
-        progress = (LockProgress) findViewById(R.id.progress);
-        rootView = (FrameLayout) findViewById(android.R.id.content);
+        setContentView(R.layout.com_auth0_activity_lock);
+        progress = (LockProgress) findViewById(R.id.com_auth0_lock_progress);
+        rootView = (FrameLayout) findViewById(R.id.com_auth0_lock_content);
 
         if (application == null) {
             fetchApplicationInfo();
