+++ lock/src/main/java/com/auth0/android/lock/utils/Application.java
@@ -1,4 +1,3 @@
-
 /*
  * Application.java
  *
