+++ lock/src/main/java/com/auth0/android/lock/utils/Strategies.java
@@ -1,4 +1,3 @@
-
 /*
  * Strategies.java
  *
