+++ lock/src/main/java/com/auth0/android/lock/utils/Strategy.java
@@ -1,4 +1,3 @@
-
 /*
  * Strategy.java
  *
