+++ lock/src/main/res/layout/com_auth0_activity_lock.xml
@@ -1,7 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout android:id="@android:id/content"
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
+    <com.auth0.android.lock.views.LockProgress
+        android:id="@+id/com_auth0_lock_progress"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true" />
+
+    <FrameLayout
+        android:id="@+id/com_auth0_lock_content"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
 </RelativeLayout>
