+++ lock/src/main/res/layout/com_auth0_activity_web_view.xml
@@ -26,7 +26,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.auth0.android.lock.net.WebViewActivity">
+    tools:context="com.auth0.android.lock.provider.WebViewActivity">
 
     <ProgressBar
         android:id="@+id/com_auth0_lock_progressbar"
