--- lock/src/main/res/layout/com_auth0_lock_activity_lock.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <com.auth0.android.lock.views.LockProgress
-        android:id="@+id/progress"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true" />
-
-    <FrameLayout
-        android:id="@android:id/content"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-</RelativeLayout>
