+++ lock/src/main/res/values/strings.xml
@@ -1,3 +1,8 @@
 <resources>
-    <string name="app_name">Lock</string>
+    <!-- Social Authentication -->
+    <string name="com_auth0_social_error_authentication">Error parsing Authentication data</string>
+    <string name="com_auth0_social_error_title">There was an error during authentication</string>
+    <string name="com_auth0_social_access_denied_message">Permissions were not granted. Try again.</string>
+    <string name="com_auth0_social_invalid_authorize_url">Auth0 authorize url not properly set.</string>
+    <string name="com_auth0_social_error_message">The application isn\'t configured properly for the social connection. Please check your Auth0\'s application configuration</string>
 </resources>
